Add rustdocs
