+++ .gitignore
@@ -1,3 +1,4 @@
 bin
 */target/
 .vscode
+package-lock.json
