+++ .travis.yml
@@ -24,7 +24,6 @@ install:
   - clang-7 --version
   - nvm install node
   - node --version
-  - ./do.sh update
 
 jobs:
   include:
