+++ ci/memo.sh
@@ -4,5 +4,7 @@ cd "$(dirname "$0")/.."
 
 set -e
 
+./do.sh update
 ./do.sh build memo
+./do.sh doc memo
 ./do.sh test memo
