+++ memo/src/lib.rs
@@ -1,1 +1,5 @@
+#![deny(missing_docs)]
+
+//! A simple program that accepts a string of encoded characters and verifies that it parses. Currently handles UTF-8.
+
 pub mod processor;
