+++ memo/src/processor.rs
@@ -1,3 +1,5 @@
+//! Program entrypoint definitions
+
 use solana_sdk::{
     account_info::AccountInfo, entrypoint, entrypoint::ProgramResult, program_error::ProgramError,
     pubkey::Pubkey,
@@ -24,18 +26,12 @@ mod tests {
         let program_id = Pubkey::new(&[0; 32]);
 
         let string = b"letters and such";
-        assert_eq!(
-            Ok(()),
-            process_instruction(&program_id, &[], string)
-        );
+        assert_eq!(Ok(()), process_instruction(&program_id, &[], string));
 
         let emoji = "🐆".as_bytes();
         let bytes = [0xF0, 0x9F, 0x90, 0x86];
         assert_eq!(emoji, bytes);
-        assert_eq!(
-            Ok(()),
-            process_instruction(&program_id, &[], &emoji)
-        );
+        assert_eq!(Ok(()), process_instruction(&program_id, &[], &emoji));
 
         let mut bad_utf8 = bytes;
         bad_utf8[3] = 0xFF; // Invalid UTF-8 byte
