+++ token/js/package.json
@@ -6,7 +6,7 @@
     "type": "git",
     "url": "https://github.com/solana-labs/solana-program-library"
   },
-  "testnetDefaultChannel": "v1.2.0",
+  "testnetDefaultChannel": "v1.2.2",
   "scripts": {
     "start": "babel-node cli/main.js",
     "lint": "npm run pretty && eslint .",
@@ -14,7 +14,7 @@
     "flow": "flow",
     "flow:watch": "watch 'flow' . --wait=1 --ignoreDirectoryPattern=/doc/",
     "lint:watch": "watch 'npm run lint:fix' . --wait=1",
-    "bpf-sdk:update": "solana-bpf-sdk-install ../../bin && npm run clean:program",
+    "bpf-sdk:update": "mkdir -p ../../bin && solana-bpf-sdk-install ../../bin && npm run clean:program",
     "build:program": "../../do.sh build token",
     "clean:program": "../../do.sh clean token",
     "test:program": "../../do.sh test token",
