+++ token/src/lib.rs
@@ -1,3 +1,8 @@
+#![deny(missing_docs)]
+
+//! An ERC20-like Token program for the Solana blockchain
+
 pub mod error;
+pub mod instruction;
 pub mod processor;
 pub mod state;
